Use a warm overlay for the image
Hi @jholetschek 
This overlay defaults to adding a dark overlay to make the text pop out. You can control it using the `overlayColor`. To disable it you write `overlayColor: transparent`.

This is just an example to show that it is possible to control the tint (you could also just disable it and just change the image instead). In this case I've added a warmer red tint as I think it suited the image. Given that the site is still work in progress I allowed myself to commit a change in this case. I hope that is okay. Just delete the line if you preferred it as it was.